Move alloc tests to their own file
Fix unused import warnings
